requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.9
Django==3.2.3
django-admin-interface==0.16.3
django-colorfield==0.4.1
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
PyFakeDados==0.0.4
Pygments==2.9.0
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
